+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Updated jasmine dependency to 3.4.0
 -   Updated MathJax config to support MathML (in addition to the original LaTeX)
 -   Updated MathJax dependency to 3.0.0
+-   Updated USGS ci/cd configuration to reflect trivial upstream changes.
+-   Removed client side login dialog. Instead show a descriptive message if the service call fails with status 401
+-   Updated to latest wdfn-viz (1.5) which uses USWDS 2.6
 
 ### Added
 -   Dockerfile and docker-compose scripts to build a pubswh-ui. This includes Jenkinsfile.build used to build the image
@@ -18,5 +21,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added additional available search parameters to the manager app's search form.
 -   Added support for sending logs to a Graylog server.
 -   Added USGS ci/cd configuration and Dockerfile-ci for building.
+-   Added Pull From SIPP button on the manager application
 
+### Fixed
+-   Tokens are refreshed are proxy service calls.
+-   publication view on pubswh now propertly returns a 404 page on non-existent publication
+-   Deletes in the manager now pass the accept headers correctly.
+-   Fixed the browse publications feature
 [Unreleased]: https://github.com/NWQMC/WQP_UI/master